requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
celery
redis
pydantic
pydantic-settings
passlib[bcrypt]
httpx
pytest
python-dotenv
pyjwt
python-multipart
